# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "KHO Minh Khai Pro V45 API",
    "version": "2.10.2",
    "description": "OpenAPI document for the route/controller/service split. JSON legacy is fallback only; business APIs should use Mongo primary routes."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/docs": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Swagger UI",
        "responses": {
          "200": {
            "description": "HTML Swagger UI"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/api/docs/openapi.json": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "OpenAPI JSON",
        "responses": {
          "200": {
            "description": "OpenAPI document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token lấy từ API đăng nhập."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token lấy từ API đăng nhập."
      }
    },
...
    "responses": {
...
      "ValidationError": {
        "description": "Dữ liệu không hợp lệ",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "success": false,
              "message": "Dữ liệu không hợp lệ",
              "code": "VALIDATION_ERROR"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Dữ liệu không hợp lệ"
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
    }
  },
...
}